requirements.txt
PyYAML==3.11
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
cracklib==2.9.3
dnspython3==1.12.0
ecdsa==0.13
kombu==3.0.26
ldap3==0.9.8.7
ocflib
paramiko==1.15.2
pexpect==3.3
pyasn1==0.1.8
pycrypto==2.6.1
pysnmp==4.2.5
pytz==2015.4
redis==2.10.3
requests==2.7.0
wheel==0.24.0
